リファクタリング (#14)
* create quiz mod

* Add expect

* fmt
