+++ src/quiz.rs
@@ -1,0 +1,3 @@
+pub mod quiz;
+
+use quiz::*;
